+++ ron-rdt/lib/RON/Data.hs
@@ -17,6 +17,7 @@ module RON.Data (
     fromRon,
     getObjectStateChunk,
     mkStateChunk,
+    newObject,
     newRon,
     objectEncoding,
     payloadEncoding,